feat(firebaseServices): Move 1KeySignNote, 2KeySigText, and 3WriteChords to new ExamQuestions folder in the componenents folder. Rename KeySigDataProps interface to UserDataProps and change in all files. Switch Decrement and Increment buttons around and rename them, turning them into Back and Forward buttons.
